moved func to lib\connections\service.go
